Move selected specimens action
