+++ study/api-src/org/labkey/api/specimen/query/SpecimenQueryView.java
@@ -17,6 +17,7 @@
 package org.labkey.api.specimen.query;
 
 import org.apache.commons.lang3.StringUtils;
+import org.jetbrains.annotations.NotNull;
 import org.labkey.api.collections.ResultSetRowMapFactory;
 import org.labkey.api.data.AbstractTableInfo;
 import org.labkey.api.data.Aggregate;
@@ -70,6 +71,7 @@ import java.io.Writer;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -423,7 +425,7 @@ public class SpecimenQueryView extends BaseSpecimenQueryView
         return createView(context, filter, createDefaultSort(viewType), viewType, false, null, false);
     }
 
-    public static SpecimenQueryView createView(ViewContext context, ParticipantDataset[] participantDatasets, ViewType viewType)
+    public static SpecimenQueryView createView(ViewContext context, @NotNull Collection<? extends ParticipantDataset> participantDatasets, ViewType viewType)
     {
         SimpleFilter filter = new SimpleFilter();
         Study study = StudyService.get().getStudy(context.getContainer());
@@ -594,10 +596,10 @@ public class SpecimenQueryView extends BaseSpecimenQueryView
         return filter;
     }
 
-    protected static SimpleFilter addFilterClause(Study study, SimpleFilter filter, ParticipantDataset[] participantDatasets)
+    protected static SimpleFilter addFilterClause(Study study, SimpleFilter filter, @NotNull Collection<? extends ParticipantDataset> participantDatasets)
     {
         boolean visitBased = study.getTimepointType().isVisitBased();
-        if (participantDatasets != null && participantDatasets.length > 0)
+        if (!participantDatasets.isEmpty())
         {
             StringBuilder whereClause = new StringBuilder();
             List<Object> params = new ArrayList<>();
